# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://backspace.webhop.me'
paths:
...
  '/api/paziente/{id}':
    get:
      summary: 'Display the specified resource.'
      operationId: displayTheSpecifiedResource
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Endpoints
      security: []
    put:
      summary: 'Update the specified resource in storage.'
      operationId: updateTheSpecifiedResourceInStorage
      description: 'Requests/UpdatePazienteRequest fa la validazione'
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: 'Must be a valid email address. Must not be greater than 255 characters.'
                  example: lroberts@example.com
                nome:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: qnntxmvmzoasnjrpvixu
                cognome:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: nuondnq
                datadinascita:
                  type: string
                  description: 'Must be a valid date.'
                  example: '2024-02-09T12:58:07'
                tipo:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: epegeajvqcuz
                medico_id:
                  type: integer
                  description: ''
                  example: 14
              required:
                - email
                - nome
                - cognome
                - datadinascita
                - tipo
      security: []
    delete:
      summary: 'Remove the specified resource from storage.'
      operationId: removeTheSpecifiedResourceFromStorage
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the paziente.'
        example: 1
        required: true
        schema:
          type: integer
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []